# docs/openapi.yml
openapi: "3.0.3"

info:
  title: udemy-reversi-webapp
  version: 1.0.0

paths:
  /api/games:
    get:
      summary: 対戦履歴を取得する
      tags: 
        - 対戦履歴を表示
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        darkMoveCount:
                          type: integer
                          example: 1
                        lightMoveCount:
                          type: integer
                          example: 1
                        winnerDisc:
                          type: integer
                          example: 1
                        startedAt:
                          type: string
                          example: 2017-07-21T17:32:28Z
                        endAt:
                          type: string
                          example: 2017-07-21T17:32:28Z
    post:
      summary: 対戦開始時の初期データを登録する
      tags:
        - 対戦する
      responses:
        201:
          description: created games

  /api/games/latest/turns/{turnCount}:
    parameters:
      - name: turnCount
        in: path
        schema:
          type: integer
        required: true
    get:
      summary: 現在（指定）の盤面を表示する
      tags:
        - 対戦する
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  turnCount:
                    type: integer
                    example: 1
                  board:
                    type: array
                    items:
                      type: array
                      items:
                        type: integer
                    example:
                      [
                        [0, 0, 0, 0, 0, 0, 0, 0],
                        [0, 0, 0, 0, 0, 0, 0, 0],
                        [0, 0, 0, 0, 0, 0, 0, 0],
                        [0, 0, 0, 1, 2, 0, 0, 0],
                        [0, 0, 0, 2, 1, 0, 0, 0],
                        [0, 0, 0, 0, 0, 0, 0, 0],
                        [0, 0, 0, 0, 0, 0, 0, 0],
                        [0, 0, 0, 0, 0, 0, 0, 0],
                      ]
                  nextDisc:
                    type: integer
                    example: 1
                  winnerDisc:
                    type: integer
                    example: 1

  /api/games/latest/turns:
    post:
      summary: プレイヤーが盤面に置いた石を登録する
      tags: 
        - 対戦する
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                turnCount:
                  type: integer
                  example: 1
                move:
                  type: object
                  properties:
                    disc:
                      type: integer
                      example: 1
                    x:
                      type: integer
                      example: 1
                    y:
                      type: integer
                      example: 1
      responses:
        201:
          description: Created
  
  /api/games/latest/turns/{turnCount}/valid-moves:
    parameters:
      - name: turnCount
        in: path
        schema:
          type: integer
        required: true
    get:
      summary: 指定したターンで有効な手の一覧を取得する
      tags:
        - 対戦する
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  validMoves:
                    type: array
                    items:
                      type: object
                      properties:
                        disc:
                          type: integer
                          example: 1
                        x:
                          type: integer
                          example: 2
                        y:
                          type: integer
                          example: 3